Added two solutions for Google Foobar Challenge 'Binary Bunnies', both in binary_bunnies.py.
solution(seq) uses a recursive binomial coefficient equation at each level of the binary tree and multiplies up.

brute_force(seq) finds all permutations of seq[1:] and recursively checks if each layer (top down) is the same and stops if they are not the same. So if layer 3 results are different it will not continue to layer 4, 5, 6, etc.